feat(boxai-sidebar): PR fixes
