+++ src/elements/content-sidebar/stories/tests/BoxAISidebar-visual.stories.tsx
@@ -1,40 +1,12 @@
 import {expect, fn, screen, userEvent, waitFor, within} from '@storybook/test';
 import { type StoryObj } from '@storybook/react';
-import {AgentType} from '@box/box-ai-agent-selector';
 import ContentSidebar from '../../ContentSidebar';
 import BoxAISidebar from '../../BoxAISidebar';
+import {mockAgents} from "../__mocks__/BoxAISidebarMocks";
 
 const mockFeatures = {
     'boxai.sidebar.enabled': true,
 };
-
-export const mockAgents: AgentType[] = [
-    {
-        id: '1',
-        name: 'Agent 1',
-        description: 'This is the default agent',
-        isEnterpriseDefault: true,
-        imageURL: 'https://cdn01.boxcdn.net/_assets/img/favicons/favicon-32x32.png',
-    },
-    {
-        id: '2',
-        name: 'Agent 2',
-        description: 'This agent has a different description',
-        isEnterpriseDefault: false,
-    },
-    {
-        id: '3',
-        name: 'Agent 3',
-        isEnterpriseDefault: false,
-        ask: { foo: 'foobar' },
-    },
-    {
-        id: '4',
-        name: 'Agent 4',
-        description: 'This is agent 4',
-        imageURL: 'https://cdn01.boxcdn.net/_assets/img/favicons/favicon-32x32.png',
-    },
-];
 
 export default {
     title: 'Elements/ContentSidebar/BoxAISidebar/tests/visual-regression-tests',
@@ -70,7 +42,7 @@ export const BoxAIWithAgentSelectorOpened: StoryObj<typeof BoxAISidebar> = {
     },
     play: async () => {
         await waitFor(async () => {
-            const agentSelector = await screen.findByRole('button', {name: 'Agent Agent 1'}, { timeout: 5000 });
+            const agentSelector = await screen.findByRole('button', {name: 'Agent Agent 1'}, { timeout: 2000 });
             expect(agentSelector).toBeInTheDocument();
 
             await userEvent.click(agentSelector);
